Fix build with Linux 6.10
During commit:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=9b163e0d330debbf7dcc14b2c3e2dc19a3b50a1d
is_dma_mapped member has been dropped since it was not used anymore. The
DMA mapping is done directly in single spi drivers in case it's supported
so having is_dma_mapped set to 1 lead to code breakage. So we can consider
it always as 0 and basically guard it if Linux version is 6.10+

Signed-off-by: Giulio Benetti <giulio.benetti@benettiengineering.com>